Fix beat ID parsing to support decimal IDs like 3.1
- Change BeatData.id from int to float
- Allows beats with IDs like 3.1, 3.2 to be loaded correctly
- Previously 3.1 was being truncated to 3, causing beat 3.1 to never load
- Fixes skipping from beat 3 directly to beat 4

+++ Velinor-Unity/Assets/Scripts/Core/BeatData.cs
@@ -8,7 +8,7 @@ using UnityEngine;
 [Serializable]
 public class BeatData
 {
-    [SerializeField] public int id;                        // 1, 2, 3, etc. or use pid for passages
+    [SerializeField] public float id;                       // 1, 2, 3, 3.1, etc. or use pid for passages
     [SerializeField] public string pid;                    // "desert_intro" (passages format) or null (beats)
     [SerializeField] public string type;                   // "player_posture", "npc_turn", "npc_shared"
     [SerializeField] public string active_speaker;         // "Player", "Nima", "Ravi", "Shared", etc.
